Refactor funcs to different files
